+++ bdtsim/renderer/game_tree.py
@@ -16,23 +16,28 @@
 # limitations under the License.
 
 import sys
+from typing import Any, Callable, Optional
 
 from graphviz import Digraph  # type: ignore
 
 from bdtsim.protocol_path import Decision, Choice
 from bdtsim.simulation_result import SimulationResult, ResultNode
-from .renderer import Renderer
+from .renderer import Renderer, ValueType
 from .renderer_manager import RendererManager
 
 
 class GameTreeRenderer(Renderer):
     def render(self, simulation_result: SimulationResult) -> None:
-        gt = SimulationGameTree(simulation_result)
+        gt = SimulationGameTree(
+            simulation_result=simulation_result,
+            autoscale_func=self.autoscale
+        )
         sys.stdout.write(gt.source + '\n')
 
 
 class SimulationGameTree(Digraph):  # type: ignore
-    def __init__(self, simulation_result: SimulationResult) -> None:
+    def __init__(self, simulation_result: SimulationResult,
+                 autoscale_func: Optional[Callable[[Any, ValueType], Any]] = None) -> None:
         super(SimulationGameTree, self).__init__(
             name='gametree',
             graph_attr={
@@ -47,6 +52,7 @@ class SimulationGameTree(Digraph):  # type: ignore
         )
 
         self._simulation_result = simulation_result
+        self._autoscale_func = autoscale_func
 
         self._generated_node_count = 0
         self._generated_subgraph_count = 0
@@ -56,6 +62,12 @@ class SimulationGameTree(Digraph):  # type: ignore
         })
 
         self._render_simulation_result()
+
+    def _autoscale(self, value: Any, value_type: ValueType) -> Any:
+        if self._autoscale_func is None:
+            return value
+        else:
+            return self._autoscale_func(value, value_type)
 
     def _render_simulation_result(self) -> None:
         self._process_node(self._simulation_result.execution_result_root)
@@ -106,11 +118,11 @@ class SimulationGameTree(Digraph):  # type: ignore
             aggregation_summary = node.aggregation_summary
             subject_summary = aggregation_summary.get(subject)
 
-            for summary_label, summary_attr in (
-                ('TX Fees', 'tx_fees'),
-                ('TX Count', 'tx_count'),
-                ('Funds Diff', 'funds_diff'),
-                ('Bal. Diff', 'balance_diff')
+            for summary_label, summary_attr, summary_type in (
+                ('TX Fees', 'tx_fees', ValueType.GAS),
+                ('TX Count', 'tx_count', ValueType.PLAIN),
+                ('Funds Diff', 'funds_diff', ValueType.WEI),
+                ('Bal. Diff', 'balance_diff', ValueType.WEI)
             ):
                 if subject_summary is None:
                     label_lines.append('<i>%s</i>: 0' % summary_label)
@@ -118,9 +130,12 @@ class SimulationGameTree(Digraph):  # type: ignore
                     value_min = getattr(subject_summary, summary_attr + '_min')
                     value_max = getattr(subject_summary, summary_attr + '_max')
                     if value_min == value_max:
-                        value_str = str(value_min)
+                        value_str = str(self._autoscale(value_min, summary_type))
                     else:
-                        value_str = '[%s, %s]' % (str(value_min), str(value_max))
+                        value_str = '[%s, %s]' % (
+                            str(self._autoscale(value_min, summary_type)),
+                            str(self._autoscale(value_max, summary_type))
+                        )
 
                     label_lines.append('<i>%s</i>: %s' % (
                         summary_label,